requirements-gradio.txt
gradio==4.26.0
geopy
autogluon.tabular==1.2
annotated-types==0.7.0
pydub==0.25.1
boto3==1.38.24
botocore==1.38.24
httpx==0.24.1
fastapi==0.101.1
starlette==0.27.0
python-dotenv
pandas
